backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
firebase-admin==6.4.0
python-dotenv==1.0.0
Pillow==10.2.0
pytest==7.4.4
pytest-asyncio==0.23.3
hypothesis==6.98.3
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
